% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{attgt_noif}
\alias{attgt_noif}
\title{attgt_noif}
\usage{
attgt_noif(attgt)
}
\arguments{
\item{attgt}{group-time average treatment effect}
}
\value{
attgt_if object
}
\description{
Class for holding group-time average treatment effects
which don't include influence functions
}
