% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_bootstrap.R
\name{attgt_pte_aggregations}
\alias{attgt_pte_aggregations}
\title{attgt_pte_aggregation}
\usage{
attgt_pte_aggregations(attgt.list, ptep)
}
\arguments{
\item{attgt.list}{list of attgt results from \code{compute.pte}}

\item{ptep}{\code{pte_params} object}
}
\value{
\code{pte_emp_boot} object
}
\description{
Aggregate group-time average treatment effects into
overall, group, and dynamic effects.  This function is only used
for computing standard errors using the empirical bootstrap.
}
