% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{group_time_att}
\alias{group_time_att}
\title{group_time_att}
\usage{
group_time_att(
  group,
  time.period,
  att,
  V_analytical,
  se,
  crit_val,
  inf_func,
  n,
  W,
  Wpval,
  alp,
  ptep
)
}
\value{
object of class \code{group_time_att}
}
\description{
group_time_att class
}
