% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{gt_data_frame}
\alias{gt_data_frame}
\title{gt_data_frame}
\usage{
gt_data_frame(data)
}
\value{
\code{gt_data_frame} object
}
\description{
function to take in a data.frame, check if it has the right
columns to be used to calculate a group-time average treatment effect,
and sets the class of the data.frame to include \code{gt_data_frame}
}
