% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_functions.R
\name{keep_all_pretreatment_subset}
\alias{keep_all_pretreatment_subset}
\title{keep_all_pretreatment_subset}
\usage{
keep_all_pretreatment_subset(data, g, tp, ...)
}
\arguments{
\item{data}{a data frame}

\item{g}{group}

\item{tp}{time period}
}
\value{
all data but in correct format for computing ATT(g,t)
}
\description{
A function that takes an original data set and keeps all
data for all groups that are not-yet-treated by period \code{tp} as well
as for group \code{g}.

In particular, this keeps more data than functions like \code{two_by_two}
subset that use a fixed base period.

A main use case for this function is the interactive fixed effects approach
proposed in Callaway and Tsyawo (2023).
}
