% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_functions.R
\name{keep_all_untreated_subset}
\alias{keep_all_untreated_subset}
\title{keep_all_untreated_subset}
\usage{
keep_all_untreated_subset(data, g, tp, ...)
}
\arguments{
\item{data}{a data frame}

\item{g}{group}

\item{tp}{time period}
}
\value{
all data but in correct format for computing ATT(g,t)
}
\description{
A function that takes an original data set and keeps all
pre-treatment data for all groups.  For group g, it also includes data
for the current period.

Also, note that if \code{tp} is still a pre-treatment period for group g,
then periods after \code{tp} will also be dropped for group g.  This is a
design choice and is useful especially for estimating placebo
group-time average treatment effects in pre-treatment periods.

A main use case for this function is to compute ATT(g,t)'s using a global
estimation strategy such as imputation in Gardner (2022).
}
