% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.PTE_bootstrap_results}
\alias{plot.PTE_bootstrap_results}
\title{Plots a summary of the bootstrap samples}
\usage{
\method{plot}{PTE_bootstrap_results}(x, ...)
}
\arguments{
\item{x}{A \code{PTE_bootstrap_results} model object built via
running the \code{PTE_bootstrap_inference} function.}

\item{...}{Other methods passed to plot}
}
\description{
Plots a summary of the bootstrap samples
}
\author{
Adam Kapelner
}
