% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.group_time_att}
\alias{print.group_time_att}
\title{print.group_time_att}
\usage{
\method{print}{group_time_att}(x, ...)
}
\arguments{
\item{x}{a \code{group_time_att} object}

\item{...}{extra arguments}
}
\description{
prints value of a \code{group_time_att} object
}
