% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.pte_results}
\alias{print.pte_results}
\title{print.pte_results}
\usage{
\method{print}{pte_results}(x, ...)
}
\arguments{
\item{x}{a \code{pte_results} object}

\item{...}{extra arguments}
}
\description{
prints value of a \code{pte_results} object
}
