% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_att_gt.R
\name{process_att_gt}
\alias{process_att_gt}
\title{process_att_gt}
\usage{
process_att_gt(att_gt_results, ptep)
}
\arguments{
\item{att_gt_results}{ATT(g,t)'s}

\item{ptep}{\code{pte_params} object}
}
\description{
process attgt results when influence function is available
}
