% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pte.R
\name{pte_default}
\alias{pte_default}
\title{pte_default}
\usage{
pte_default(
  yname,
  gname,
  tname,
  idname,
  data,
  xformla = ~1,
  d_outcome = FALSE,
  d_covs_formula = ~-1,
  lagged_outcome_cov = FALSE,
  est_method = "dr",
  anticipation = 0,
  base_period = "varying",
  cband = TRUE,
  alp = 0.05,
  boot_type = "multiplier",
  biters = 100,
  cl = 1
)
}
\arguments{
\item{yname}{Name of outcome in \code{data}}

\item{gname}{Name of group in \code{data}}

\item{tname}{Name of time period in \code{data}}

\item{idname}{Name of id in \code{data}}

\item{data}{balanced panel data}

\item{xformla}{one-sided formula for covariates used in the propensity score
and outcome regression models}

\item{d_outcome}{Whether or not to take the first difference of the outcome.
The default is FALSE.  To use difference-in-differences, set this to be TRUE.}

\item{d_covs_formula}{A formula for time varying covariates to enter the
first estimation step models.  The default is not to include any, and, hence,
to only include pre-treatment covariats.}

\item{lagged_outcome_cov}{Whether to include the lagged outcome as a covariate.
Default is FALSE.}

\item{est_method}{Which type of estimation method to use. Default is "dr" for
doubly robust.  The other option is "reg" for regression adjustment.}

\item{alp}{significance level; default is 0.05}

\item{boot_type}{should be one of "multiplier" (the default) or "empirical".
The multiplier bootstrap is generally much faster, but \code{attgt_fun} needs
to provide an expression for the influence function (which could be challenging
to figure out).  If no influence function is provided, then the \code{pte}
package will use the empirical bootstrap no matter what the value of this
parameter.}

\item{biters}{number of bootstrap iterations; default is 100}

\item{cl}{number of clusters to be used when bootstrapping; default is 1}
}
\value{
\code{pte_results} object
}
\description{
This is a generic/example wrapper for a call to the \code{pte} function.

This function provides access to difference-in-differences and unconfoundedness
based identification/estimation strategies given (i) panel data and (ii)
staggered treatment adoption
}
