% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_bootstrap.R
\name{qott_pte_aggregations}
\alias{qott_pte_aggregations}
\title{qott_pte_aggregations}
\usage{
qott_pte_aggregations(attgt.list, ptep, extra_gt_returns)
}
\arguments{
\item{attgt.list}{list of attgt results from \code{compute.pte}}

\item{ptep}{\code{pte_params} object}
}
\value{
\code{pte_emp_boot} object
}
\description{
Aggregate group-time distribution of the treatment effect into
overall, group, and dynamic effects.
}
