% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{summary.group_time_att}
\alias{summary.group_time_att}
\title{summary.group_time_att}
\usage{
\method{summary}{group_time_att}(object, ...)
}
\arguments{
\item{object}{an \code{group_time_att} object}

\item{...}{extra arguments}
}
\description{
prints a summary of a \code{group_time_att} object
}
