% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_and_summarize.R
\name{summary.PTE_bootstrap_results}
\alias{summary.PTE_bootstrap_results}
\title{Prints a summary of the model to the console}
\usage{
\method{summary}{PTE_bootstrap_results}(object, ...)
}
\arguments{
\item{object}{A \code{PTE_bootstrap_results} model object built via
running the \code{PTE_bootstrap_inference} function.}

\item{...}{Other methods passed to summary}
}
\description{
Prints a summary of the model to the console
}
\examples{
library(PTE)
data(continuous_example)
pte_results = PTE_bootstrap_inference(continuous_example$X, continuous_example$y,
    regression_type = "continuous", B = 5, num_cores = 1)
summary(pte_results)

}
\author{
Adam Kapelner
}
