% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{summary.pte_emp_boot}
\alias{summary.pte_emp_boot}
\title{summary.pte_emp_boot}
\usage{
\method{summary}{pte_emp_boot}(object, ...)
}
\arguments{
\item{object}{a \code{pte_emp_boot} object}

\item{...}{additional function arguments}
}
\value{
\code{summary.pte_results} object
}
\description{
Summary for \code{pte_emp_boot} object
}
