% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{summary.pte_results}
\alias{summary.pte_results}
\title{summary.pte_results}
\usage{
\method{summary}{pte_results}(object, ...)
}
\arguments{
\item{object}{an \code{pte_results} object}

\item{...}{other arguments}
}
\description{
A function to summarize \code{pte} results.
}
