% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pte_params.R
\name{t2orig}
\alias{t2orig}
\title{t2orig}
\usage{
t2orig(t, original_time.periods)
}
\arguments{
\item{t}{a particular time period to convert back to original time
periods.}

\item{original_time.periods}{vector containing all original time periods.}
}
\value{
original time period converted from new time period
}
\description{
A helper function to switch from "new" t values to
original t values.  This allows for periods not being exactly spaced
apart by 1.
}
