% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{attgt_if}
\alias{attgt_if}
\title{attgt_if}
\usage{
attgt_if(attgt, inf_func)
}
\arguments{
\item{attgt}{group-time average treatment effect}

\item{inf_func}{influence function}
}
\value{
attgt_if object
}
\description{
Class for holding group-time average treatment effects
along with their influence function
}
