% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTE.R
\docType{data}
\name{continuous_example}
\alias{continuous_example}
\title{Mock RCT data with a continuous endpoint.}
\description{
A list with two objects (a) \code{X}, a dataframe with n rows representing clinical subjects and columns: 
treatment, x1, x2, x3, x4 and x5 where treatment is binary indicating the two arms of the clinical trial
and x1, ..., x5 are covariates that were collected about each subject and (b) \code{y}, a length n vector storing 
the continuous response values where, in this mock dataset, larger values indicate "better" outcomes for the 
subjects.
}
\author{
My Name \email{kapelner@qc.cuny.edu}
}

