% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{group_time_att}
\alias{group_time_att}
\title{group_time_att}
\usage{
group_time_att(
  group,
  time.period,
  att,
  V_analytical,
  se,
  crit_val,
  inf_func,
  n,
  W,
  Wpval,
  cband,
  alp,
  ptep,
  extra_gt_returns
)
}
\arguments{
\item{group}{numeric vector of groups for ATT(g,t)}

\item{time.period}{numeric vector of time periods for ATT(g,t)}

\item{att}{numeric vector containing the value of ATT(g,t) for
corresponding group and time period}

\item{V_analytical}{analytical asymptotic variance matrix for ATT(g,t)'s}

\item{se}{numeric vector of standard errors}

\item{crit_val}{critical value (usually a critical value for conducting
uniform inference)}

\item{inf_func}{matrix of influence function}

\item{n}{number of unique individuals}

\item{W}{Wald statistic for ATT(g,t) version of pre-test of parallel trends
assumption}

\item{Wpval}{p-value for Wald pre-test of ATT(g,t) version of parallel
trends assumption}

\item{cband}{whether or not to report a uniform (instead of pointwise)
confidence band (default is TRUE)}

\item{alp}{significance level}

\item{ptep}{\code{pte_params} object}

\item{extra_gt_returns}{A place to return anything extra from particular
group-time average treatment effect calculations.  For DID, this might
be something like propensity score estimates, regressions of untreated
potential outcomes on covariates.  For ife, this could be something
like the first step regression 2sls estimates.  This argument is also
potentially useful for debugging.}
}
\value{
object of class \code{group_time_att}
}
\description{
group_time_att class
}
