% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_and_summarize.R
\name{plot.PTE_bootstrap_results}
\alias{plot.PTE_bootstrap_results}
\title{Plots a summary of the bootstrap samples}
\usage{
\method{plot}{PTE_bootstrap_results}(x, ...)
}
\arguments{
\item{x}{A \code{PTE_bootstrap_results} model object built via
running the \code{PTE_bootstrap_inference} function.}

\item{...}{Other methods passed to plot}
}
\description{
Plots a summary of the bootstrap samples
}
\examples{
library(PTE)
data(continuous_example)
pte_results = PTE_bootstrap_inference(continuous_example$X, continuous_example$y,
    regression_type = "continuous", B = 5, num_cores = 1)
plot(pte_results)

}
\author{
Adam Kapelner
}
