% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{print.PTE_bootstrap_results}
\alias{print.PTE_bootstrap_results}
\title{Prints a summary of the model to the console}
\usage{
\method{print}{PTE_bootstrap_results}(x, ...)
}
\arguments{
\item{x}{A \code{PTE_bootstrap_results} model object built via
running the \code{PTE_bootstrap_inference} function.}

\item{...}{Other methods passed to print}
}
\description{
Prints a summary of the model to the console
}
\author{
Adam Kapelner
}
