% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pte.R
\name{pte}
\alias{pte}
\title{pte}
\usage{
pte(
  yname,
  gname,
  tname,
  idname,
  data,
  subset_fun,
  attgt_fun,
  alp = 0.05,
  biters = 100,
  cl = 1,
  ...
)
}
\arguments{
\item{yname}{Name of outcome in \code{data}}

\item{gname}{Name of group in \code{data}}

\item{tname}{Name of time period in \code{data}}

\item{idname}{Name of id in \code{data}}

\item{attgt_fun}{This is a function that should work in the case where
there is a single group and the "right" number of time periods to
recover an estimate of the ATT.  For example, in the contest of
difference in differences, it would need to work for a single group,
find the appropriate comparison group (untreated units), find the right
time periods (pre- and post-treatment), and then recover an estimate
of ATT for that group.  It will be called over and over separately
by groups and by time periods to compute ATT(g,t)'s.

The function needs to work in a very specific way.  It should take in the
arguments: \code{data}, \code{group}, \code{time.periods}, \code{glist},
\code{tlist}, and \code{...}.  \code{...} are additional arguments (such as
formulas for covariates) that \code{attgt_fun} needs.  From these arguments
\code{attgt_fun} must return a list with element \code{ATT} containing the
group-time average treatment effect for that group and that time period.

If \code{attgt_fun} returns an influence function (which should be provided
in a list element named \code{inf_func}), then the code will use the
multiplier bootstrap to compute standard errors for group-time average
treatment effects, an overall treatment effect parameter, and a dynamic
treatment effect parameter (i.e., event study parameter).  If
\code{attgt_fun} does not return an influence function, then the same
objects will be computed using the empirical bootstrap.  This is usually
(perhaps substantially) easier to code, but also will usually be (perhaps
substantially) computationally slower.}

\item{alp}{significance level; default is 0.05}

\item{biters}{number of bootstrap iterations; default is 100}

\item{cl}{number of clusters to be used when bootstrapping; default is 1}
}
\description{
Main function for computing panel treatment effects
}
