% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTE.R
\docType{package}
\name{PTE}
\alias{PTE-package}
\alias{PTE}
\title{Personalized Medicine Inference}
\description{
Personalized Medicine...
}
\references{
Kapelner, A, Bleich, J, Cohen, ZD, DeRubeis, RJ and Berk, R (2014) Inference for Treatment Regime Models in Personalized Medicine, arXiv
}
\author{
Adam Kapelner \email{kapelner@qc.cuny.edu}, Alina Levine and Justin Bleich
}
\keyword{Personalized}
\keyword{bootstrap}
\keyword{medicine,}
