% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTE.R
\docType{package}
\name{PTE}
\alias{PTE-package}
\alias{PTE}
\title{Personalized Medicine Inference}
\description{
Personalized Medicine...
}
\references{
Kapelner, A, Bleich, J, Cohen, ZD, DeRubeis, RJ and Berk, R (2021) Inference for Treatment Regime Models in Personalized Medicine. Frontiers in Big Data. \url{https://www.frontiersin.org/journals/big-data/articles/10.3389/fdata.2021.572532/full}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kapelner/PTE}
  \item Report bugs at \url{https://github.com/kapelner/PTE/issues}
}

}
\author{
Adam Kapelner \email{kapelner@qc.cuny.edu}, Alina Levine and Justin Bleich
}
\keyword{bootstrap}
\keyword{medicine}
\keyword{personalized}
