% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_inference.R
\name{PTE_bootstrap_inference}
\alias{PTE_bootstrap_inference}
\title{Bootstrap inference for a prespecified personalization / recommendation model}
\usage{
PTE_bootstrap_inference(X, y, regression_type = "continuous",
  personalized_model_build_function = NULL, censored = NULL,
  predict_function = function() {     predict(mod, obs_left_out) },
  cleanup_mod_function = NULL, y_higher_is_better = TRUE, verbose = TRUE,
  full_verbose = FALSE, H_0_mu_equals = 0, pct_leave_out = 0.1,
  m_prop = 1, B = 3000, alpha = 0.05, plot = TRUE, num_cores = 1)
}
\arguments{
\item{X}{A \eqn{n \times p}{n x p} dataframe of covariates where one column is labeled "treatment" and it
is a binary vector of treatment allocations in the study.}

\item{y}{An \eqn{n}-length numeric vector which is the response}

\item{regression_type}{A string indicating the regression problem. Legal values are "continous" (the response \code{y} is
a real number with no missing data, the default), "incidence" (the reponse \code{y} is
either 0 or 1) and "survival". If the type is "survival", the user must also supply additional data via the 
parameter \code{censored}.}

\item{personalized_model_build_function}{An R function that will be evaluated to construct the personalized medicine / recommendation 
model. In the formula for the model, the response is "y", the treatment vector is 
"treatment" and the data is "Xyleft". This function must return some type of object
that can be used for prediction later via \code{predict_function}.}

\item{censored}{Only required if the \code{regression_type} is "survival". In this case, this vector is of length \eqn{n} and is binary 
   where 0 indicates censored and 1 indicates uncensored. In a clinical trial, someone who is still alive 
   at the end of the study or was lost to follow up will receive a censor value of 0, while someone who died during the study 
will receive a censor value of 1. 
   \eqn{n} and is binary where 0 indicates censorship (e.g. the patient died).}

\item{predict_function}{An R cfunction that will be evaluated on left out data after the model is built with the training data. This function
uses the object "mod" that is the result of the \code{personalized_model_build_function} and it must make use of
"obs_left_out", a vector of covariates representing an observation from \code{X}. This function must return a 
scalar numeric quantity for comparison. The default function is \code{predict(mod, obs_left_out)}.}

\item{cleanup_mod_function}{A function that is called at the end of a cross validation iteration to cleanup the model 
in some way. This is used for instance if you would like to release the memory your model is using but generally does not apply.
The default is \code{NA} for "no function."}

\item{y_higher_is_better}{True if a response value being higher is clinically "better" than one that is lower (e.g. cognitive ability in a drug trial for the 
mentally ill). False if the response value being lower is clinically "better" than one that is higher (e.g. amount of weight lost 
in a weight-loss trial). Default is \code{TRUE}.}

\item{verbose}{Prints out a dot for each bootstrap sample. This only works on some platforms.}

\item{full_verbose}{Prints out full information for each cross validation model for each bootstrap sample. This only works on some platforms.}

\item{H_0_mu_equals}{The \eqn{\mu_{I_0}}{mu_I_0} value in \eqn{H_0}{H_0}. Default is 0 which answers the question: does my allocation procedure do better than a naive
allocation procedure.}

\item{pct_leave_out}{In the cross-validation, the proportion of the original dataset left out to estimate out-of-sample metrics. The default is 0.1
which corresponds to 10-fold cross validation.}

\item{m_prop}{Within each bootstrap sample, the proportion of the total number of rows of \code{X} to sample without replacement. \code{m_prop < 1} ensures
the number of rows sampled is less than \code{n} which fixes the consistency of the bootstrap estimator of a non-smooth functional. The default 
is 1 since non-smoothness may not be a common issue.}

\item{B}{The number of bootstrap samples to take. We recommend making this as high as you can tolerate given speed considerations.
The default is 3000.}

\item{alpha}{Defines the confidence interval size (1 - alpha). Defaults to 0.05.}

\item{plot}{Illustrates the estimate, the bootstrap samples and the confidence intervals on a histogram plot. Default to TRUE.}

\item{num_cores}{The number of cores to use in parallel to run the bootstrap samples more rapidly. Defaults to serial by using 1 core.}
}
\value{

}
\description{
Runs B bootstrap samples using a prespecified model then computes the two I estimates based on cross validation. 
p values of the two I estimates are computed for a given \eqn{H_0: \mu_{I_0} = \mu_0}{H_0: mu_I_0 = mu_0} and 
confidence intervals are provided.
}
\author{
Adam Kapelner
}
