% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{pte_emp_boot}
\alias{pte_emp_boot}
\title{pte_emp_boot}
\usage{
pte_emp_boot(
  attgt_results,
  overall_results,
  group_results,
  dyn_results,
  extra_gt_returns = NULL
)
}
\arguments{
\item{attgt_results}{\code{data.frame} holding attgt results}

\item{overall_results}{\code{data.frame} holding overall results}

\item{group_results}{\code{data.frame} holding group results}

\item{dyn_results}{\code{data.frame} holding dynamic results}
}
\value{
pte_emp_boot object
}
\description{
pte_emp_boot

Class for holding \code{pte} empirical bootstrap results
}
