% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pte_params.R
\name{pte_params}
\alias{pte_params}
\title{pte_params}
\usage{
pte_params(yname, gname, tname, idname, data, glist, tlist, alp, biters, cl)
}
\arguments{
\item{yname}{Name of outcome in \code{data}}

\item{gname}{Name of group in \code{data}}

\item{tname}{Name of time period in \code{data}}

\item{idname}{Name of id in \code{data}}

\item{glist}{list of groups to create group-time average treatment effects
for}

\item{tlist}{list of time periods to create group-time average treatment
effects for}

\item{alp}{significance level; default is 0.05}

\item{biters}{number of bootstrap iterations; default is 100}

\item{cl}{number of clusters to be used when bootstrapping; default is 1}
}
\description{
pte_params
}
