% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_bootstrap.R
\name{qtt_pte_aggregations}
\alias{qtt_pte_aggregations}
\title{qtt_pte_aggregations}
\usage{
qtt_pte_aggregations(attgt.list, ptep, extra_gt_returns)
}
\arguments{
\item{attgt.list}{list of attgt results from \code{compute.pte}}

\item{ptep}{\code{pte_params} object}

\item{extra_gt_returns}{A place to return anything extra from particular
group-time average treatment effect calculations.  For DID, this might
be something like propensity score estimates, regressions of untreated
potential outcomes on covariates.  For ife, this could be something
like the first step regression 2sls estimates.  This argument is also
potentially useful for debugging.}
}
\value{
\code{pte_emp_boot} object
}
\description{
Aggregate group-time distributions into qtt versions of
overall, group, and dynamic effects.
}
