% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pte_params.R
\name{setup_pte}
\alias{setup_pte}
\title{Setup panel treatment effects parameters}
\usage{
setup_pte(
  yname,
  gname,
  tname,
  idname,
  data,
  required_pre_periods = 1,
  anticipation = 0,
  base_period = "varying",
  cband = TRUE,
  alp = 0.05,
  boot_type = "multiplier",
  weightsname = NULL,
  gt_type = "att",
  ret_quantile = 0.5,
  biters = 100,
  cl = 1,
  ...
)
}
\value{
\code{pte_params} object
}
\description{
This is a function for how to setup
the data to be used in the \code{pte} package.

The \code{setup_pte} function builds on \code{setup_pte_basic} and
attempts to provide a general purpose function (with error handling)
to arrange the data in a way that can be processed by \code{subset_fun}
and \code{attgt_fun} in the next steps.
}
