% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pte_params.R
\name{setup_pte}
\alias{setup_pte}
\title{setup_pte}
\usage{
setup_pte(yname, gname, tname, idname, data, alp = 0.05, biters = 100, cl = 1)
}
\arguments{
\item{yname}{Name of outcome in \code{data}}

\item{gname}{Name of group in \code{data}}

\item{tname}{Name of time period in \code{data}}

\item{idname}{Name of id in \code{data}}

\item{data}{balanced panel data}

\item{alp}{significance level; default is 0.05}

\item{biters}{number of bootstrap iterations; default is 100}

\item{cl}{number of clusters to be used when bootstrapping; default is 1}
}
\description{
Setup panel treatment effects parameters
}
