% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_and_summary.R
\name{summary.PTE_bootstrap_results}
\alias{summary.PTE_bootstrap_results}
\title{Prints a summary of the model to the console}
\usage{
\method{summary}{PTE_bootstrap_results}(object, ...)
}
\arguments{
\item{object}{A \code{PTE_bootstrap_results} model object built via
running the \code{PTE_bootstrap_inference} function.}

\item{...}{Other methods passed to summary}
}
\description{
Prints a summary of the model to the console
}
\author{
Adam Kapelner
}
