% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_functions.R
\name{two_by_two_subset}
\alias{two_by_two_subset}
\title{two_by_two_subset}
\usage{
two_by_two_subset(data, g, tp, control_group = "notyettreated", ...)
}
\description{
A function for computing a 2x2 subset of original data.
This is the subset with post treatment periods separately for the
treated group and comparison group and pre-treatment periods in the period
immediately before the treated group became treated.
}
